# Conflicts:
#	src/A_cal_salary.java
#	src/Calculate.java
แก้ไขข้อผิดพลาดจากการ merge brach เป็นเวอร์ชั่นปัจจุบัน

+++ src/Calculate.java
@@ -1,10 +1,10 @@
 
 public class Calculate{
 	private String name;
-	private int salary;
-	private int sale;
+	private double salary;
+	private double sale;
 	
-	public Calculate (String name,int salary,int sale){
+	public Calculate (String name,double salary,double sale){
 		this.name = name;
 		this.salary = salary;
 		this.sale = sale;
